+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Internal refactoring of submit button
+
 # 0.11.3
 
 - Enhance error message when attachment reflection is missing